+++ src/Providers/AbstractCompanyProvider.php
@@ -2,10 +2,12 @@
 namespace Uniondrug\PolicySdk\Providers;
 
 use Uniondrug\PolicySdk\Injectable;
+use Uniondrug\PolicySdk\InjectableTrait;
 use Uniondrug\PolicySdk\Structs\Config;
 
 abstract class AbstractCompanyProvider extends Injectable
 {
+    use InjectableTrait;
     /**
      * 保司的相关配置
      * @Config
